Make UpdateCaseOnly concurrency safe in ServiceHearingsService with feature toggle
