+++ build.gradle
@@ -340,7 +340,7 @@ dependencies {
   implementation group: 'ch.qos.logback', name: 'logback-core', version: '1.2.10'
   implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.10'
 
-  implementation group: 'com.github.hmcts', name: 'sscs-common', version: '5.3.1'
+  implementation group: 'com.github.hmcts', name: 'sscs-common', version: '5.3.6'
 
   implementation group: 'org.slf4j', name: 'slf4j-api', version: '1.7.36'
