+++ charts/sscs-hearings-api/values.yaml
@@ -54,3 +54,4 @@ java:
     SSCS_SERVICE_CODE: BBA3
     ELINKS_V2_FEATURE_ENABLED: false
     HMC_DEPLOYMENT_FILTER_ENABLED: false
+    SERVICE_HEARING_SERVICE_V2_ENABLED: false
